requirements.txt
praw==7.8.1
python-dotenv==1.0.0
numpy>=1.21.0,<2.0
pandas==2.0.3
spotipy==2.22.1
openai>=1.0.0
flask>=2.3.0
flask-cors>=4.0.0